add test and fix circular dependency
