+++ k8s/so-rest-deployment.yaml
@@ -19,7 +19,7 @@ spec:
       - name: ci-tno-dots
       containers:
       - name: so-rest # container name
-        image: ghcr.io/ees-tue/simulation-orchestrator:4.0.0
+        image: ghcr.io/ees-tue/simulation-orchestrator:5.0.0
         imagePullPolicy: Always
         ports:
         - containerPort: 8001
