+++ simulation_orchestrator/parse_esdl.py
@@ -5,7 +5,7 @@ from esdl import esdl, EnergySystem
 from esdl.esdl_handler import EnergySystemHandler
 from base64 import b64decode
 
-from simulation_orchestrator.rest.schemas.simulation_schemas import CalculationService
+from simulation_orchestrator.rest.schemas.CalculationService import CalculationService
 from simulation_orchestrator.dataclasses.dataclasses import CalculationServiceInfo
 from simulation_orchestrator.models.model_inventory import Model
 from simulation_orchestrator.types import EsdlId, ProgressState
